vsr: liveness, nack prepares from before log view
A DVC with a log-view `v` includes _some_ of the
potentially committed messages for views `≥v`.

However, it will include all potentially committed
messages from views `<v`:

log-view=v means that a replica accepted an SV
message for that view (or the replica was a
primary for that view). In any case, the starting
headers for that view were durably persisted by
the replica.

So all those headers must be a part of DVC the
replica subsequently sends.

This means that if

- we are deciding whether we should truncate op=x
- we _don't_ know the canonical header for op=x
- and we have an uncanonical DVC with a header for op=x with view<log_view_canonical

That DVC is a nack for `op=x`. We don't know which
header would be canonical, but it definitely isn't
the one from the DVC.

Seed: 15677965580699030051
Closes: #1510